requirements.txt
streamlit
python-dotenv
pypdf2
reportlab
groq
matplotlib
pandas
